+++ CHANGELOG.md
@@ -27,7 +27,7 @@ BUG FIXES:
 - Fix check validity of secret at create response (if invalid, the response will not be stored in blockchain).
 - Fix wrong callback result for create identity.
 - Fix duplicate AS data signature processing and set data received on RP side.
-- Fix bug when as_id_list is empty.
+- Fix bug when `as_id_list` is not provided when creating a request.
 
 ## 0.10.2 (October 29, 2018)